+++ backend/src/main/java/com/agrupae/infrastructure/persistence/jpa/model/user/UserJpaEntity.java
@@ -1,0 +1,44 @@
+package com.agrupae.infrastructure.persistence.jpa.model.user;
+
+import java.time.Instant;
+import java.util.UUID;
+
+import com.agrupae.domain.role.Role;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Table(name = "users")
+@Entity
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Setter(value = AccessLevel.PROTECTED)
+public class UserJpaEntity {
+    @Id
+    private UUID id;
+    @Column(nullable = false)
+    private String name;
+    @Column(nullable = false, unique = true)
+    private String email;
+    @Column(nullable = false)
+    private String passwordHash;
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Role role;
+    @Column(nullable = false)
+    private Instant createdAt;
+    @Column(nullable = false)
+    private Instant updatedAt;
+}